Continue refactor. Fix bug with provisional tracks.
Note this might bring back #4728.
