+++ src/openrct2/world/entrance.h
@@ -33,6 +33,7 @@ void game_command_remove_park_entrance(sint32 *eax, sint32 *ebx, sint32 *ecx, si
 
 typedef struct rct_xyz16;
 typedef struct rct_xyzd16;
+typedef struct rct_map_element;
 
 extern bool gParkEntranceGhostExists;
 extern rct_xyz16 gParkEntranceGhostPosition;
@@ -43,8 +44,13 @@ extern money32 gParkEntranceGhostPrice;
 
 extern rct_xyzd16 gParkEntrances[MAX_PARK_ENTRANCES];
 
+extern rct_xyzd16 gRideEntranceExitGhostPosition;
+extern uint8 gRideEntranceExitGhostStationIndex;
+
 void park_entrance_remove_ghost();
 money32 park_entrance_place_ghost(sint32 x, sint32 y, sint32 z, sint32 direction);
 void reset_park_entrance();
+void maze_entrance_hedge_replacement(sint32 x, sint32 y, rct_map_element *mapElement);
+void maze_entrance_hedge_removal(sint32 x, sint32 y, rct_map_element *mapElement);
 
 #endif
